requirements.txt
Flask
Flask-SQLAlchemy
stripe
python-dotenv